function-app/requirements.txt
azure-functions
azure-identity
azure-storage-blob
azure-eventhub